update master (#3)
* Add "roles/orgpolicy.policyViewer" to get rid of permission warnings/errors when running inventory.

* Never prompt during Forseti instance upgrades

When the Forseti client and server instances were running updates on
startup, it was possible for apt/dpkg to prompt for input during the
upgrade process. Since there's no tty attached to the associated process
this would cause the startup script to hang.

This commit modifies the invocation of `apt-get update` to never prompt,
update default conffiles, and use the original file when a non-default
file is present.