Continue server installation when config-validator policy is absent
The config-validator installation process fetches policy from the GCS
bucket, but this policy is not available by default. In this default
case the server startup script will call `gsutil cp` to sync down the
policy, fail, and then the `set -e` will cause the entire installation
process to abort.

This commit fixes the issue by allowing the `gsutil` command to fail. In
the case of this failure we print a message indicating that the sync was
skipped and continue on. The config-validator is in alpha and is allowed
to fail at this point, so bypassing this step is acceptable.