Strip out Forseti installation from real time enforcer example
The Forseti client and server are currently independent from the
real-time-enforcer module; creating and testing these resources
duplicates the test behavior of the simple-example module and increases
the surface area for failure. In addition the GCP organization IAM
policies in the Forseti module tend to trigger race conditions when all
suites are run at once, so this test coverage is actively harmful.

This commit strips out the Forseti installation from the real time
enforcer. Users can combine the modules together if desired, but
starting with decoupled functionality will help users make an
intentional decision when deciding how they want to run Forseti and the
real time enforcer.