Add limited retries to real time enforcer services
We've observed random but fairly common CI failures with the
real-time-enforcer module, specifically in that the enforcer service is
not running. Further inspection into one of these failures showed that
the `docker pull` failed due to a network error and the service shut
down. Network issues are unavoidable so we should be able to handle
these situations with some resilience.

This commit updates the opa-server and enforcer systemd units to restart
the services upon error, delay 5 seconds between each attempt, and to
stop restarting if 5 starts occur within 120 seconds. The delay between
each restart allows transient network issues to resolve, and the start
limit burst prevents the service from endlessly restarting in case
there's an issue that can't be recovered from through restarting.