Support composite root resources
Currently the Forseti module supports a single root resource, specified
in the server configuration file as `root_resource_id`. Forseti now
supports multiple root resources, but we don't expose this functionality
in this module.

This commit adds a `composite_root_resources` variable for specifying
multiple roots for Forseti to scan. Composite root resources will be
preferred, and the module will fall back to the singular root resource
when composite root resources are not present.